src/test/java/ru/yandex/practicum/filmorate/repository/jdbc/JdbcFilmRepositoryTest.java
package ru.yandex.practicum.filmorate.repository.jdbc;

import lombok.RequiredArgsConstructor;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.jdbc.Sql;
import ru.yandex.practicum.filmorate.model.Film;
import ru.yandex.practicum.filmorate.model.User;

import java.util.List;

import static org.assertj.core.api.Assertions.assertThat;
import static org.assertj.core.api.Assertions.assertThatList;
import static org.junit.jupiter.api.Assertions.assertEquals;

@SpringBootTest
@AutoConfigureTestDatabase
@RequiredArgsConstructor(onConstructor_ = @Autowired)
@Sql(scripts = "classpath:TestFilmData.sql")
class JdbcFilmRepositoryTest {
    private final JdbcFilmRepository filmRepository;

    @Test
    void shouldGetAllFilms() {
        List<Film> filmsList = (List<Film>) filmRepository.getAll();
        assertThatList(filmsList).hasSizeBetween(4, 4);
    }
}